gkrp_data_portal/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gkrp-data-portal"
version = "0.1.0"
description = "A data portal for entry an analysis for the GKRP project."
readme = ""
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Lyubomir Angelov", email = "lyubomir.angelov.an@gmail.com" }
]

dependencies = [
    "alembic==1.18.1",
    "loguru==0.7.3,",
    "nicegui-3.5.0",
    "fastapi==0.128.0",
]

# This tells installers where the importable code actually lives.
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["gkrp_data_portal*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
shared_library = []